+++ shuup/admin/modules/orders/views/edit.py
@@ -340,6 +340,7 @@ class OrderEditView(CreateOrUpdateView):
                 "name": customer.full_name,
                 "phone_no": customer.phone,
                 "email": customer.email,
+                "tax_number": getattr(customer, "tax_number", ""),
                 "companies": [force_text(company) for company in companies] if len(companies) else None,
                 "groups": [force_text(group) for group in customer.groups.all()],
                 "merchant_notes": customer.merchant_notes